Create Maximum Possible Sum
Bi≥max(Ai,Ai+1), this condition is most important.
For maximum sum, the end values are :
a[0]=b[0] & a[n-1]=b[n-2]

Then a[i] is inter-related with both b[i] & b[i-1], so for staisfying Bi≥max(Ai,Ai+1) & having max possible sum,a[i] should be minimum of b[i] & b[i-1].
After updating all the values in array, we take the sum & output it. 
This algorithm is O(n) complexity.